refactor: 인수 테스트 패키지 변경
